requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
chardet==4.0.0
discord.py==1.7.3
flake8==3.9.2
idna==3.2
idna-ssl==1.1.0
importlib-metadata==4.6.4
mccabe==0.6.1
multidict==5.1.0
mypy==0.910
mypy-extensions==0.4.3
pkg-resources==0.0.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
yarl==1.6.3
zipp==3.5.0